move key pair generation to util module
